requirements.txt
django-social-auth
django-appengine-toolkit